refactor: remove unnecessary components, integrate activities, branding and shopping cart views
